src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
langchain==0.0.335
langchain-community==0.0.10
langchain-openai==0.0.5
supabase==1.2.0
python-dotenv==1.0.0
spacy==3.7.2
numpy==1.26.1
pandas==2.1.3
httpx==0.25.1
pytest==7.4.3
redis==5.0.1
PyJWT==2.8.0
requests==2.31.0 